fix loc of DBSpendUpdateTransactions
